Remove activate subcommand, move code out of `cli` package
