--- cli/errors.go
@@ -1,5 +1,0 @@
-package cli
-
-import "fmt"
-
-var errStepNotAvailableOfflineMode error = fmt.Errorf("step not available in offline mode")
